pyproject.toml
[tool.poetry]
name = "flask-base"
version = "0.1.0"
description = "flask web server rapidly develop template."
authors = ["dv008 <dv008@secret.com>"]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.8"
flask-restx = "^0.2.0"
flask-CORS = "^3.0.9"
gunicorn = "^20.0.4"
Flask-SQLAlchemy = "^2.4.4"
Flask-bcrypt = "^0.7.1"
marshmallow = "^3.9.1"
loguru = "^0.5.3"
flask-jwt-extended = "^3.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
